requirements.txt

phidata
streamlit
youtube_transcript_api
pytube
openai==0.28